requirements.txt
ultralytics>=8.0.0
kaggle>=1.5.13
pyyaml>=6.0
tqdm>=4.64.0
